\name{time-package}
\alias{time-package}
\alias{time}
\docType{package}
\title{
\packageTitle{time}
}
\description{
\packageDescription{time}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{time}
\packageIndices{time}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{time}

Maintainer: \packageMaintainer{time}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
