% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_phenotype.R
\name{time_phenotype}
\alias{time_phenotype}
\title{time_phenotype}
\usage{
time_phenotype(data = data)
}
\arguments{
\item{data}{The matrix of gene expression data, rows are samples and columns are genes.}
}
\value{
a list
}
\description{
Get the TIME phenotypes and TIMEscore
}
\details{
Use gene expression data to get the corresponding TIME phenotypes and TIMEscore
}
\examples{
if(!require("GSVA")) BiocManager::install("GSVA",update = F,ask = F)
if(!require("clusterRepro")) install.packages("clusterRepro",update = F,ask = F)
library(TIME)
data('data')
result=time_phenotype(data2)
}
